pyproject.toml
[project]
name = "task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
